survey画面デザイン
